feat: mediaPlayer 에 repeatMode, shuffle 여부 추가, duration long으로 변경 (#14)
* feat: playerStatus 에 RepeatMode, shuffle 여부 필드 추가

* feat: track duration int -> long으로 변경

* feat: player 에 필요한 repeatMode, shuffle 여부 변경 동작 2개 추가

+++ core/datasource/src/main/java/com/hsikkk/delightroom/datasource/media/LocalMediaDataSourceImpl.kt
@@ -109,7 +109,7 @@ class LocalMediaDataSourceImpl(
                     Track(
                         id = id,
                         name = cursor.getString(nameColumn),
-                        duration = cursor.getInt(durationColumn),
+                        duration = cursor.getLong(durationColumn),
                         albumId = cursor.getLong(albumIdColumn),
                         albumName = cursor.getString(albumColumn),
                         artist = cursor.getString(albumArtistColumn),
